refactor: 한 클래스에서만 사용되는 GameResultViewScope 클래스 이동
